-Added TEControllers -Added spelling fix -Removed extra widget -Added Code fix
