--- lib/screens/user_registration/widgets/signup_button.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SignUpButton extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return RaisedButton(
-      onPressed: null,
-      child: Text('Sign Up'),
-    );
-  }
-}
